pyproject.toml
[project]
name = "caddie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cadquery-ocp>=7.8.1.1.post1",
    "ipython>=9.6.0",
    "jupyter>=1.1.1",
    "neo4j>=5.28.2",
    "openai-agents>=0.3.2",
    "pymupdf>=1.26.4",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
]
